+++ defaults/controllers/_default/root/admin/workers/servers/_index.js
@@ -1,0 +1,12 @@
+// --------------------
+// Workers-Servers join resource controller
+// Route definition
+// --------------------
+
+// exports
+
+// route definition
+exports = module.exports = {
+	// route params
+	type: 'resourceJoin'
+};